Add mock server response for an invalid Submit Health Event request. Move mock server interactions from mockConfig.js to test/support/default.js so that they can be customized for different service calls
