+++ test/support/defaults.js
@@ -1,6 +1,10 @@
 const { mock } = require('pactum');
 const { Before, BeforeAll, AfterAll } = require('@cucumber/cucumber');
+const { like } = require('pactum-matchers');
 const mockConfig = require('./mockConfig');
+
+const sheResponse = require('../fixtures/Bundle-DDCC-TX-SHE-response-example-1.json');
+const sheResponseError = require('../fixtures/OperationOutcome-example.json');
 
 BeforeAll(() => {
   // perform some shared setup
@@ -14,6 +18,44 @@ Before(function (testCase, callback) {
       .then(() => callback())
       .catch((error) => callback(error))
   }
+});
+
+Before({ tags: '@SubmitHealthEvent' }, () => {
+  mock.addInteraction([{
+    id: 'submitHealthEvent',
+    strict: false,
+    request: {
+      method: 'POST',
+      path: '/submitHealthEvent',
+      body: {
+        resourceType: like('Bundle'),
+        id: like('DDCC-TX-SHE-bundle-example-1')
+      }
+    },
+    response: {
+      status: 201,
+      body: sheResponse
+    }
+  }]);
+});
+
+Before({ tags: '@SubmitInvalidHealthEvent' }, () => {
+  mock.addInteraction([{
+    id: 'submitHealthEvent',
+    strict: false,
+    request: {
+      method: 'POST',
+      path: '/submitHealthEvent',
+      body: {
+        resourceType: like('Bundle'),
+        id: like('DDCC-TX-SHE-bundle-example-invalid')
+      }
+    },
+    response: {
+      status: 422,
+      body: sheResponseError
+    }
+  }]);
 });
 
 AfterAll(() => {
